Updated minimumDifficulty To BigInteger

+++ src/main/java/co/rsk/federate/signing/hsm/message/PowHSMBlockchainParameters.java
@@ -2,15 +2,17 @@ package co.rsk.federate.signing.hsm.message;
 
 import co.rsk.crypto.Keccak256;
 
+import java.math.BigInteger;
+
 /**
  * Created by Kelvin Isievwore on 09/05/2023.
  */
 public class PowHSMBlockchainParameters {
     private final Keccak256 checkpoint;
-    private final int minimumDifficulty;
+    private final BigInteger minimumDifficulty;
     private final String network;
 
-    public PowHSMBlockchainParameters(String checkpoint, int minimumDifficulty, String network) {
+    public PowHSMBlockchainParameters(String checkpoint, BigInteger minimumDifficulty, String network) {
         this.checkpoint = new Keccak256(checkpoint);
         this.minimumDifficulty = minimumDifficulty;
         this.network = network;
@@ -20,7 +22,7 @@ public class PowHSMBlockchainParameters {
         return checkpoint;
     }
 
-    public int getMinimumDifficulty() {
+    public BigInteger getMinimumDifficulty() {
         return minimumDifficulty;
     }
